Don't recreate closed PRs for the same project/version pair
For each version update, check for an existing closed PR and skip
create-pull-request if found.  This allows suppressing semiweekly branch
update emails for broken upstream versions that will never merge.  We
retain the branch so the PR can be manually reopened, after which we'll
refresh it on the next autoupdate run.